% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_LNfiles.R
\name{check_LNfiles}
\alias{check_LNfiles}
\title{Check LexisNexis TXT files}
\usage{
check_LNfiles(x, encoding = "UTF-8", verbose = TRUE)
}
\arguments{
\item{x}{Name or names of LexisNexis TXT file to be converted.}

\item{encoding}{Encoding to be assumed for input files. Defaults to UTF-8 (the LexisNexis standard value).}

\item{verbose}{A logical flag indicating whether information should be printed to the screen.}
}
\description{
Read a LexisNexis TXT file and check consistency.
}
\details{
Can check consistency of LexisNexis txt files. read_LN needs at least Beginning, End and length in each article to work
}
\examples{
Track the time
}
\author{
Johannes B. Gruber
}
\keyword{LexisNexis}
