% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisNexisTools.R
\name{lnt2bibtex}
\alias{lnt2bibtex}
\title{Convert LNToutput to other formats}
\usage{
lnt2bibtex(x, art_id, ...)
}
\arguments{
\item{x}{An object of class LNToutput.}

\item{art_id}{The ID(s) of the article(s) to convert.}

\item{...}{unused.}
}
\description{
Takes output from \link{lnt_read} and converts chosen articles to a bibtex citation.
}
\examples{
LNToutput <- lnt_read(lnt_sample())

docs <- lnt_convert(LNToutput, art_id = 1)
}
