% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisNexisTools.R
\name{lnt_add}
\alias{lnt_add}
\title{Adds or replaces articles}
\usage{
lnt_add(to, what, where = "meta", replace = TRUE)
}
\arguments{
\item{to}{A LNToutput object to which something should be added.}

\item{what}{A data.frame which is added.}

\item{where}{Either "meta", "articles" or "paragraphs" to indicate the slot to which data is added.}

\item{replace}{If TRUE, will overwrite entries which have the same ID as}
}
\description{
This functions adds a dataframe to a slot in an LNToutput object or overwrite
existing entries. The main use of the function is to add an extract of one of
the data.frames back to an LNToutput object after operations were performed
on it.
}
\examples{
lnt_sample()
# LNToutput <- lnt_add(to = LNToutput, what = ups, where = "meta", replace = TRUE)
}
\author{
Johannes Gruber
}
