% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisNexisTools.R
\name{lnt_lookup}
\alias{lnt_lookup}
\title{Lookup keywords in articles}
\usage{
lnt_lookup(x, pattern, cluster = NULL, case_insensitive = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{A string or vector of strings or an LNToutput object.}

\item{pattern}{A character vector of keywords. Word boundaries before and
after the keywords are honoured. Regular expression can be used.}

\item{cluster}{The number of CPU cores to use. Use \code{NULL} or \code{1} to
turn off.}

\item{case_insensitive}{If FALSE, the pattern matching is case sensitive and if
TRUE, case is ignored during matching.}

\item{verbose}{A logical flag indicating whether a status bar is printed to
the screen.}
}
\value{
A list keyword hits.
}
\description{
This function looks in for the provided keywords in the
  provided string or LNToutput object. This can be useful, for example, to
  see which of the keywords you used when retrieving the data was used in
  each article.
}
\details{
If an LNToutput object is provided, the function will look for the
  pattern in the headlines and articles. The returned object is a list of
  hits. If a regular expression is provided, the retunred word will be the
  actual value from the text.
}
\examples{
# Make LNToutput object from sample
LNToutput <- lnt_read(lnt_sample())

# Lookup keywords
LNToutput@meta$Keyword <- lnt_lookup(LNToutput,
                                     "statistical computing")

}
\author{
Johannes Gruber
}
