% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisNexisTools.R
\docType{class}
\name{LNToutput-class}
\alias{LNToutput-class}
\title{An S4 class to store the three data.frames created with \link{lnt_read}.}
\description{
This S4 class stores the output from \link{lnt_read}. Objects consist of
three data.frames which you can select using \code{@}.
}
\section{Slots}{

\describe{
\item{\code{meta}}{The metadata of the articles read in.}

\item{\code{articles}}{The article texts and respective IDs.}

\item{\code{paragraphs}}{The paragraphs (if the data.frame exists) and respective
article and paragraph IDs.}
}}

