% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisNexisTools.R
\docType{methods}
\name{LNToutput_methods}
\alias{LNToutput_methods}
\alias{show,LNToutput-method}
\alias{[,LNToutput,ANY,ANY,ANY-method}
\title{Methods for LNToutput output objects}
\usage{
\S4method{show}{LNToutput}(object)

\S4method{[}{LNToutput,ANY,ANY,ANY}(x, i, j, invert = FALSE)
}
\arguments{
\item{x, object}{An LNToutput object.}

\item{i}{Rows of the meta data.frame (default) or values of j.}

\item{j}{The a column you want to use to subset the LNToutput object. Takes
character strings.}

\item{invert}{Invert the selection of i.}
}
\description{
Methods for LNToutput output objects
}
