% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexisNexisTools.R
\name{lnt_checkFiles}
\alias{lnt_checkFiles}
\title{Check LexisNexis TXT files (deprecated)}
\usage{
lnt_checkFiles(...)
}
\arguments{
\item{...}{No functionality as this was deprecated.}
}
\description{
Check LexisNexis TXT files (deprecated)
}

