% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnt_convert.R
\name{lnt_convert}
\alias{lnt_convert}
\alias{lnt2rDNA}
\title{Convert LNToutput to other formats}
\usage{
lnt_convert(x, to = "rDNA", what = "Articles")

lnt2rDNA(x, what)
}
\arguments{
\item{x}{An object of class LNToutput}

\item{to}{Which format to convert into.}

\item{what}{Either "Articles" or "Paragraph" to use articles or paragraphs as
text in the output object.}
}
\description{
Takes output from \link{lnt_read} and converts it to other formats.
}
\examples{
LNToutput <- lnt_read(lnt_sample())
docs <- lnt_convert(LNToutput)
}
