% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnt_rename.R
\name{lnt_rename}
\alias{lnt_rename}
\title{Assign proper names to LexisNexis TXT files}
\usage{
lnt_rename(x, encoding = "UTF-8", recursive = FALSE, report = FALSE,
  simulate = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{Can be either a character vector of LexisNexis TXT file name(s),
folder name(s) or can be left blank (see example).}

\item{encoding}{Encoding to be assumed for input files. Defaults to UTF-8
(the LexisNexis standard value).}

\item{recursive}{A logical flag indicating whether subdirectories are
searched for more txt files.}

\item{report}{A logical flag indicating whether the function will return a
report which files were renamed.}

\item{simulate}{Should the renaming be simulated instead of actually done?
This can help prevent accidental renaming of unrelated txt files which
happen to be in the same directory as the files from 'LexisNexsis'.}

\item{verbose}{A logical flag indicating whether information should be
printed to the screen.}
}
\description{
Give proper names to LN TXT files based on search term and period retrieved
from each file cover page. This information is not always delivered by
LexisNexis though. If the information is not present in the file, new file
names will be empty.
}
\details{
Warning: This will rename all txt files in a give folder.
}
\examples{

# Copy sample file to current wd
lnt_sample()

# Rename files in current wd and report back if successful
 \dontrun{report.df <- lnt_rename(recursive = FALSE,
                        report = TRUE)}

# Or provide file name(s)
my_files<-list.files(pattern = ".txt", full.names = TRUE,
                     recursive = TRUE, ignore.case = TRUE)
report.df <- lnt_rename(x = my_files,
                        recursive = FALSE,
                        report = TRUE)

# Or provide folder name(s)
report.df <- lnt_rename(x = getwd())

report.df
}
\author{
Johannes B. Gruber
}
\keyword{LexisNexis}
