% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnt_similarity.R
\name{lnt_similarity}
\alias{lnt_similarity}
\title{Check for highly similar articles.}
\usage{
lnt_similarity(texts, dates, IDs = NULL, threshold = 0.99,
  Rel.diff.on = FALSE)
}
\arguments{
\item{texts}{column with articles.}

\item{dates}{column with dates.}

\item{IDs}{IDs of articles.}

\item{threshold}{At which threshold of similarity is an article considered a
duplicate.}

\item{Rel.diff.on}{Calculate the relative Levenstein distance between two
articles if set to TRUE (can take very long). The main difference between
the similarity and distance value is that the distance takes word order
into account while similarity employs the bag of words approach.}
}
\description{
Check for highly similar articles by comparing all articles published on the
same date.
}
\details{
The function produces a data.frame consisting of information about
  duplicated articles
}
\examples{
\dontrun{
duplicates.df <- lnt_similarity(texts = LNToutput@articles$Article,
                               dates = LNToutput@meta$Date,
                               IDs = LNToutput@articles$ID)
meta.df <- LNToutput@meta[!LNToutput@meta$ID \%in\% 
                           duplicates.df$ID.duplicate,]
articles.df <- LNToutput@articles[!LNToutput@articles$ID \%in\% 
                                   duplicates.df$ID.duplicate,]
paragraphs.df <- LNToutput@paragraphs[!LNToutput@paragraphs$ID \%in\% 
                                      duplicates.df$ID.duplicate,]
                                      }
}
\author{
Johannes B. Gruber
}
\keyword{similarity}
