% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_LN.R
\name{rename_LNfiles}
\alias{rename_LNfiles}
\title{Assign proper names to LexisNexis TXT files}
\usage{
rename_LNfiles(x, encoding = "UTF-8", recursive = TRUE, report = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{Name or names of LexisNexis TXT file to be renamed or a folder location. If folder location is provided, string must end with /}

\item{encoding}{Encoding to be assumed for input files. Defaults to UTF-8 (the LexisNexis standard value).}

\item{recursive}{A logical flag indicating whether sub directories are searched for mor txt files}

\item{report}{A logical flag indicating whether the function will return a report which files were renamed}

\item{verbose}{A logical flag indicating whether information should be printed to the screen.}
}
\description{
Give proper names to LN txt files based on search term and period retrieved from each files cover page.
}
\details{
Can check consistency of LexisNexis txt files. read_LN needs at least Beginning, End and length in each article to work
}
\examples{
report.df <- rename_LNfiles("C:/Test/", recursive = TRUE, report = TRUE)
Track the time
}
\author{
Johannes B. Gruber
}
\keyword{LexisNexis}
